Add first cut at 'voice-web' application
This is a simple web-based application that allows users to
annotate human voice segments in audio data.

One good thing about this application is that it runs entirely
inside the browser -- you just need to serve static files to
host the application.

A big shout out to katspaugh, whose wavesurfer.js enabled this
app to exist.

Signed-off-by: Fujimoto Seiji <fujimoto@ceptord.net>